Getting read of useless class wrapper
