+++ project/srcs/main.cpp
@@ -6,12 +6,12 @@
 /*   By: scarboni <scarboni@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/19 22:03:00 by scarboni          #+#    #+#             */
-/*   Updated: 2022/03/21 18:02:31 by scarboni         ###   ########.fr       */
+/*   Updated: 2022/03/24 10:28:45 by scarboni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/webserv.h"
-#include "Config.hpp"
+#include "Node.hpp"
 
 int		main(int ac, char *av[])
 {
@@ -21,10 +21,10 @@ int		main(int ac, char *av[])
 		return 0;
 	}
 
-	Config *config = Config::factory(av[1]);
-	std::cout << *config << std::endl;
-	if(!config)
+	Node *firstNode = Node::digestConfigurationFile(av[1]);
+	std::cout << *firstNode << std::endl;
+	if(!firstNode)
 		return 0;
-	delete config;
+	delete firstNode;
 	return 0;
 }
